+++ CHANGELOG.md
@@ -4,7 +4,11 @@
 
 - coming soon...
 
-## 0.0.6 (2020-xx-xx Unreleased)
+## 0.0.7 (2020-xx-xx Unreleased)
+
+- coming soon...
+
+## 0.0.6 (2020-01-07)
 
 - run selected build
 - make context for build